feat: depreciate onBlurValue prop
With recent changes, "onBlurValue" seems obsolete. User onBlur is
called, so there's no real reason to keep "onBlurValue"

Breaking Change: "onBlurValue" no longer supported
