+++ src/components/__tests__/CurrencyInput-precision.spec.tsx
@@ -5,7 +5,7 @@ import CurrencyInput from '../CurrencyInput';
 const id = 'validationCustom01';
 
 describe('<CurrencyInput /> component > decimalScale', () => {
-  const onBlurValueSpy = jest.fn();
+  const onValueChangeSpy = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -13,21 +13,31 @@ describe('<CurrencyInput /> component > decimalScale', () => {
 
   it('should pad to decimalScale of 5 on blur', () => {
     const view = shallow(
-      <CurrencyInput id={id} prefix="£" onBlurValue={onBlurValueSpy} decimalScale={5} />
+      <CurrencyInput id={id} prefix="£" onValueChange={onValueChangeSpy} decimalScale={5} />
     );
     view.find(`#${id}`).simulate('blur', { target: { value: '£1.5' } });
-    expect(onBlurValueSpy).toBeCalledWith('1.50000', undefined);
+    expect(onValueChangeSpy).toBeCalledWith('1.50000', undefined);
 
     const updatedView = view.update();
     expect(updatedView.find(`#${id}`).prop('value')).toBe('£1.50000');
   });
 
   it('should pad to decimalScale of 2 on blur', () => {
+    const onBlurSpy = jest.fn();
     const view = shallow(
-      <CurrencyInput id={id} prefix="£" onBlurValue={onBlurValueSpy} decimalScale={2} />
+      <CurrencyInput
+        id={id}
+        prefix="£"
+        onValueChange={onValueChangeSpy}
+        onBlur={onBlurSpy}
+        decimalScale={2}
+      />
     );
-    view.find(`#${id}`).simulate('blur', { target: { value: '£1' } });
-    expect(onBlurValueSpy).toBeCalledWith('1.00', undefined);
+    const event = { target: { value: '£1' } };
+
+    view.find(`#${id}`).simulate('blur', event);
+    expect(onBlurSpy).toBeCalledWith(event);
+    expect(onValueChangeSpy).toBeCalledWith('1.00', undefined);
 
     const updatedView = view.update();
     expect(updatedView.find(`#${id}`).prop('value')).toBe('£1.00');
